Move grammars to grammars directory
