+++ .gitattributes
@@ -1,8 +1,12 @@
-/interface/src/parser.c linguist-generated
-/interface/src/grammar.json linguist-generated
-/interface/src/node-types.json linguist-generated
-/interface/src/tree_sitter/* linguist-vendored
-/ocaml/src/parser.c linguist-generated
-/ocaml/src/grammar.json linguist-generated
-/ocaml/src/node-types.json linguist-generated
-/ocaml/src/tree_sitter/* linguist-vendored
+* text eol=lf
+
+grammars/interface/src/*.json linguist-generated
+grammars/interface/src/parser.c linguist-generated
+
+grammars/ocaml/src/*.json linguist-generated
+grammars/ocaml/src/parser.c linguist-generated
+
+include/tree_sitter/* linguist-vendored
+
+bindings/** linguist-generated
+binding.gyp linguist-generated
