+++ .gitignore
@@ -13,9 +13,9 @@ Package.resolved
 
 # These files would be generated by 'tree-sitter generate' with the default
 # settings. We don't want them because there's already a copy at the root.
-/interface/Cargo.toml
-/interface/binding.gyp
-/interface/bindings
-/ocaml/Cargo.toml
-/ocaml/binding.gyp
-/ocaml/bindings
+grammars/interface/Cargo.toml
+grammars/interface/binding.gyp
+grammars/interface/bindings
+grammars/ocaml/Cargo.toml
+grammars/ocaml/binding.gyp
+grammars/ocaml/bindings
