+++ Cargo.toml
@@ -12,12 +12,14 @@ edition = "2018"
 
 build = "bindings/rust/build.rs"
 include = [
-  "common/*",
   "bindings/rust/*",
-  "ocaml/grammar.js",
-  "ocaml/src/*",
-  "interface/grammar.js",
-  "interface/src/*",
+  "grammars/ocaml/grammar.js",
+  "grammars/ocaml/src/*.c",
+  "grammars/ocaml/src/*.json",
+  "grammars/interface/grammar.js",
+  "grammars/interface/src/*.c",
+  "grammars/interface/src/*.json",
+  "include/*",
   "queries/*"
 ]
