+++ Package.swift
@@ -15,16 +15,17 @@ let package = Package(
             name: "TreeSitterOCaml",
             path: ".",
             sources: [
-                "interface/src/parser.c",
-                "interface/src/scanner.c",
-                "ocaml/src/parser.c",
-                "ocaml/src/scanner.c",
+                "grammars/interface/src/parser.c",
+                "grammars/interface/src/scanner.c",
+                "grammars/ocaml/src/parser.c",
+                "grammars/ocaml/src/scanner.c",
+                "include",
             ],
             resources: [
                 .copy("queries")
             ],
             publicHeadersPath: "bindings/swift",
-            cSettings: [.headerSearchPath("ocaml/src")]
+            cSettings: [.headerSearchPath("include")]
         ),
         .testTarget(
             name: "TreeSitterOCamlTests",
