+++ bindings/rust/build.rs
@@ -1,14 +1,18 @@
 fn main() {
     let root_dir = std::path::Path::new(".");
-    let ocaml_dir = root_dir.join("ocaml").join("src");
-    let interface_dir = root_dir.join("interface").join("src");
+    let include_dir = root_dir.join("include");
+    let grammars_dir = root_dir.join("grammars");
+    let ocaml_dir = grammars_dir.join("ocaml").join("src");
+    let interface_dir = grammars_dir.join("interface").join("src");
 
     let mut c_config = cc::Build::new();
-    c_config.include(&ocaml_dir);
+    c_config.include(&include_dir);
     c_config
         .flag_if_supported("-Wno-unused-parameter")
         .flag_if_supported("-Wno-unused-but-set-variable")
         .flag_if_supported("-Wno-trigraphs");
+
+    println!("cargo:rerun-if-changed={}", include_dir.to_str().unwrap());
 
     for dir in &[ocaml_dir, interface_dir] {
         let parser_path = dir.join("parser.c");
