+++ grammars/interface/corpus/module-items.txt
@@ -1,0 +1,1 @@
+../../ocaml/corpus/module-items.txt
