+++ grammars/interface/grammar.js
@@ -1,0 +1,7 @@
+module.exports = grammar(require('../ocaml/grammar'), {
+  name: 'ocaml_interface',
+
+  rules: {
+    compilation_unit: $ => optional($._signature)
+  }
+})
