+++ grammars/ocaml/package.json
@@ -1,0 +1,3 @@
+{
+  "main": "../../bindings/node/ocaml"
+}
