+++ grammars/ocaml/src/tree_sitter
@@ -1,0 +1,1 @@
+../../../include/tree_sitter
