--- interface/corpus/module-items.txt
@@ -1,1 +1,0 @@
-../../ocaml/corpus/module-items.txt
