--- interface/package.json
@@ -1,3 +1,0 @@
-{
-  "main": "../bindings/node/interface"
-}
