--- ocaml/corpus/comments.txt
@@ -1,79 +1,0 @@
-==============================
-Comments
-==============================
-
-(**)
-(* *)
-(**
-  doc
-*)
-( *)
-
----
-
-(compilation_unit
-  (comment)
-  (comment)
-  (comment)
-  (expression_item (value_path (parenthesized_operator (mult_operator)))))
-
-==============================
-Strings and comments
-==============================
-
-(* "*)" *)
-"(**)";;
-"(*";;
-
----
-
-(compilation_unit
-  (comment)
-  (expression_item (string (string_content)))
-  (expression_item (string (string_content))))
-
-==============================
-Characters and comments
-==============================
-
-(* '"' *)
-(* f' '"' *)
-(*  *)
-(* '\o170' '"' *)
-(* '\'"' *)
-(* '\" " *)
-(* '" " *)
-(* '*)
-x
-
----
-
-(compilation_unit
-  (comment)
-  (comment)
-  (comment)
-  (comment)
-  (comment)
-  (comment)
-  (comment)
-  (comment)
-  (expression_item (value_path (value_name))))
-
-==============================
-Quoted strings and comments
-==============================
-
-(* {|*)|} *)
-(* {%id |*)|} *)
-(* {%M.id|*)|} *)
-{|(**)|};;
-{|(*|};;
-
----
-
-(compilation_unit
-  (comment)
-  (comment)
-  (comment)
-  (expression_item (quoted_string (quoted_string_content)))
-  (expression_item (quoted_string (quoted_string_content))))
