--- ocaml/corpus/toplevel.txt
@@ -1,31 +1,0 @@
-==============================
-Toplevel directives
-==============================
-
-#help
-#use "file.ml"
-#show_module M
-#show x
-
----
-
-(compilation_unit
-  (toplevel_directive (directive))
-  (toplevel_directive (directive) (string (string_content)))
-  (toplevel_directive (directive) (module_path (module_name)))
-  (toplevel_directive
-    (directive)
-    (value_path (value_name))))
-
-==============================
-Line number directives
-==============================
-
-# 1 "test.ml"
-# 2 "test.ml" (* comment *)
-
----
-
-(compilation_unit
-  (line_number_directive)
-  (line_number_directive))
